Implemented a file reader
Helicopter data is now read in from a file. This file is located in the
assets/Data folder. There is also a spreadsheet file for easier data
manipulation.
Argument input for Helicopter changed slightly. The position is now
calculated from altitude and the x and z coord are set to 0.
The classes Helicopter and helicopterType now also contain toString
methods for testing purposes.
Also removed unused imports from Helicopter3D.